backend/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.1
Flask-Migrate>=3.0.0
Flask-CORS>=3.0.10
werkzeug>=2.0.0
psycopg2-binary>=2.9.1
Flask-JWT-Extended